+++ lib/nba_rb/player/defense_tracking.rb
@@ -1,0 +1,7 @@
+module NbaRb
+  module Player
+    class DefenseTracking < PlayerDashboard
+      @endpoint = 'playerdashptshotdefend'
+    end
+  end
+end
