+++ lib/nba_rb/player/pass_tracking.rb
@@ -1,0 +1,15 @@
+module NbaRb
+  module Player
+    class PassTracking < PlayerDashboard
+      @endpoint = 'playerdashptpass'
+
+      def passes_made
+        create_stats_hash(@data[0])
+      end
+
+      def passes_recieved
+        create_stats_hash(@data[1])
+      end
+    end
+  end
+end
