+++ lib/nba_rb/player/player_profile.rb
@@ -1,0 +1,19 @@
+module NbaRb
+  module Player
+    class Profile < PlayerCareer
+      @endpoint = 'playerprofilev2'
+
+      def season_highs
+        create_stats_hash(@data[10])
+      end
+
+      def career_highs
+        create_stats_hash(@data[11])
+      end
+
+      def next_game
+        create_stats_hash(@data[12])
+      end
+    end
+  end
+end
