+++ lib/nba_rb/player/rebound_log_tracking.rb
@@ -1,0 +1,7 @@
+module NbaRb
+  module Player
+    class ReboundLogTracking < PlayerDashboard
+      @endpoint = 'playerdashptreboundlogs'
+    end
+  end
+end
