+++ lib/nba_rb/player/shot_log_tracking.rb
@@ -1,0 +1,7 @@
+module NbaRb
+  module Player
+    class ShotLogTracking < PlayerDashboard
+      @endpoint = 'playerdashptshotlog'
+    end
+  end
+end
